refactor: 소모임 가입 패키지 분리
- #20
