+++ src/main/java/com/example/wegather/groupJoin/domain/vo/JoinRequestStatus.java
@@ -1,4 +1,4 @@
-package com.example.wegather.group.domain.vo;
+package com.example.wegather.groupJoin.domain.vo;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
