+++ src/main/java/com/example/wegather/groupJoin/dto/GroupJoinRequestDto.java
@@ -1,6 +1,6 @@
-package com.example.wegather.group.dto;
+package com.example.wegather.groupJoin.dto;
 
-import com.example.wegather.group.domain.entity.SmallGroupJoin;
+import com.example.wegather.groupJoin.domain.entity.SmallGroupJoin;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
